+++ tool/build.ts
@@ -6,7 +6,7 @@ const banner = `// ==UserScript==
 // @name:ko        vim comic viewer
 // @description    Universal comic reader
 // @description:ko 만화 뷰어 라이브러리
-// @version        9.0.0
+// @version        9.0.1
 // @namespace      https://greasyfork.org/en/users/713014-nanikit
 // @exclude        *
 // @match          http://unused-field.space/