+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+### 9.0.1
+
+- Restore default background color.
+- Increase contrast of scrollbar.
+
 # 9.0.0
 
+- Add the following for configuration memory.
+  // @grant GM_setValue
+  // @grant GM_getValue
 - Remove `setGmXhr`, use `setTampermonkeyApi` instead.
 - Add background color configuration.
 - Remember preceding single page count.